Move PortfolioTabManager into file
